+++ CHANGELOG.md
@@ -1,17 +1,39 @@
 # Code Editor Package for Visual Studio
 
+## [2.0.11] - 2021-07-01
+
+Integration:
+
+- Added support for Visual Studio and Visual Studio For Mac 2022.
+- Fixed an issue when the package was enabled for background processes.
+
+Project generation:
+
+- Use absolute paths for Analyzers and rulesets.
+
+
+## [2.0.10] - 2021-06-10
+
+Project generation:
+
+- Improved project generation performance when a file is moved, deleted or modified.
+
+Integration:
+
+- Improved Inner-loop performance by avoiding to call the package manager when looking up `vswhere` utility.
+- Fixed a network issue preventing the communication between Visual Studio and Unity on Windows.
+
 ## [2.0.9] - 2021-05-04
 
 Project generation:
 
-Added support for CLI.
+- Added support for CLI.
 
 Integration:
 
-Improved performance when discovering Visual Studio installations.
-Warn when legacy assemblies are present in the project.
-Warn when the package version is not up-to-date.
-
+- Improved performance when discovering Visual Studio installations.
+- Warn when legacy assemblies are present in the project.
+- Warn when the package version is not up-to-date.
 
 ## [2.0.8] - 2021-04-09